rename poke URLs from conversations (plural) to conversation (singular)
