+++ front/lib/utils/url-to-poke.ts
@@ -12,15 +12,15 @@ const ROUTE_MAPPINGS: RouteMapping[] = [
   // Conversations (old /assistant/, /agent/, and new /conversation/ URLs)
   {
     pattern: /^\/w\/([^/]+)\/assistant\/([^/]+)$/,
-    pokePath: "/poke/$1/conversations/$2",
+    pokePath: "/poke/$1/conversation/$2",
   },
   {
     pattern: /^\/w\/([^/]+)\/agent\/([^/]+)$/,
-    pokePath: "/poke/$1/conversations/$2",
+    pokePath: "/poke/$1/conversation/$2",
   },
   {
     pattern: /^\/w\/([^/]+)\/conversation\/([^/]+)$/,
-    pokePath: "/poke/$1/conversations/$2",
+    pokePath: "/poke/$1/conversation/$2",
   },
 
   // Assistants
@@ -201,7 +201,7 @@ export function convertPokeToUrl(pokeUrl: string): string | null {
 
     // Define reverse mappings
     const reverseMappings: Array<[RegExp, string]> = [
-      [/^([^/]+)\/conversations\/([^/]+)$/, "/w/$1/conversation/$2"],
+      [/^([^/]+)\/conversation\/([^/]+)$/, "/w/$1/conversation/$2"],
       [/^([^/]+)\/agents\/([^/]+)$/, "/w/$1/builder/agents/$2"],
       [/^([^/]+)\/spaces\/([^/]+)\/apps\/([^/]+)$/, "/w/$1/spaces/$2/apps/$3"],
       [
