+++ front/pages/api/poke/workspaces/[wId]/conversations/[cId]/index.ts
@@ -4,7 +4,7 @@ import { apiErrorForConversation } from "@app/lib/api/assistant/conversation/hel
 import { withSessionAuthenticationForPoke } from "@app/lib/api/auth_wrappers";
 import { Authenticator } from "@app/lib/auth";
 import type { SessionWithUser } from "@app/lib/iam/provider";
-import { getPokeConversation } from "@app/lib/poke/conversations";
+import { getPokeConversation } from "@app/lib/poke/conversation";
 import { apiError } from "@app/logger/withlogging";
 import type { PokeConversationType, WithAPIErrorResponse } from "@app/types";
